requirements.txt
asgiref==3.5.0
dj-database-url==0.5.0
Django==2.2
django-cors-headers==3.10.0
django-debug-toolbar==1.11.1
django-heroku==0.3.1
django-utils-six==2.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
gunicorn==20.1.0
psycopg2==2.9.1
psycopg2-binary==2.8.5
PyJWT==2.1.0
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
typing_extensions==4.0.1
whitenoise==3.3.1
